+++ lib/widgets/boolean_setting.dart
@@ -9,7 +9,8 @@ class BooleanSetting extends StatefulWidget {
   BooleanSetting({
     required this.setting,
     required this.settingText,
-    this.onChanged,
+    this.customOnChanged,
+    this.doSetState = true,
     super.key,
   });
 
@@ -19,7 +20,14 @@ class BooleanSetting extends StatefulWidget {
 
   /// The text shown to the user explaining the setting.
   final String settingText;
-  final Function(bool value)? onChanged;
+
+  /// Allows the implementation of additional actions beyond just changing the value
+  final Function(bool value)? customOnChanged;
+
+  /// If set to false then the BooleanSetting will not update when its value is changed.
+  /// To make it update when doSetState is set to false you need to update it in a parent widget or use a builder.
+  final bool doSetState;
+
   @override
   _BooleanSettingState createState() => _BooleanSettingState();
 }
@@ -29,17 +37,23 @@ class _BooleanSettingState extends State<BooleanSetting> {
 
   late final String settingText;
   late final Function(bool value)? onChanged;
+  late final bool doSetState;
 
   @override
   void initState() {
     super.initState();
+    doSetState = widget.doSetState;
     settingText = widget.settingText;
-    onChanged =
-        (widget.onChanged ??
-                (final bool value) {
-                  widget.setting = value;
-                })
-            as Function(bool value)?;
+    onChanged = (final bool value) {
+      (widget.customOnChanged ?? (_) {})(value);
+      if (doSetState) {
+        setState(() {
+          widget.setting = value;
+        });
+      } else {
+        widget.setting = value;
+      }
+    };
   }
 
   @override
@@ -68,90 +82,6 @@ class _BooleanSettingState extends State<BooleanSetting> {
                 ),
                 value: widget.setting,
                 onChanged: onChanged,
-              ),
-            ),
-          ),
-
-          SizedBox(
-            width: MediaQuery.of(context).size.width * 0.6,
-            child: Text(
-              settingText,
-              style: TextStyle(
-                color: Settings.colorSet.text,
-                fontSize: SizeConfiguration.settingInfoText,
-                decoration: TextDecoration.none,
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-/// This version does not automate the setting handling
-class LightBooleanSetting extends StatefulWidget {
-  const LightBooleanSetting({
-    required this.settingText,
-    required this.onChanged,
-    required this.initValue,
-    super.key,
-  });
-
-  /// The text shown to the user explaining the setting.
-  final String settingText;
-  final bool initValue;
-  final Function(bool value) onChanged;
-  @override
-  _LightBooleanSettingState createState() => _LightBooleanSettingState();
-}
-
-class _LightBooleanSettingState extends State<LightBooleanSetting> {
-  _LightBooleanSettingState();
-
-  late final String settingText;
-  late final Function(bool value) onChanged;
-  late bool value;
-
-  @override
-  void initState() {
-    super.initState();
-    settingText = widget.settingText;
-    onChanged = widget.onChanged;
-    value = widget.initValue;
-  }
-
-  @override
-  Widget build(final BuildContext context) {
-    return Container(
-      margin: const EdgeInsets.only(top: 40),
-      color: Settings.colorSet.primary,
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: <Widget>[
-          Container(
-            margin: EdgeInsets.only(
-              right: MediaQuery.of(context).size.width * 0.1,
-            ),
-            child: Material(
-              color: Colors.transparent,
-              child: Switch(
-                activeColor: Settings.colorSet.tertiary,
-                inactiveThumbColor: Settings.colorSet.secondary,
-                inactiveTrackColor: Settings.colorSet.inactiveTrack,
-                trackOutlineColor: WidgetStateProperty.resolveWith(
-                  (final Set<WidgetState> states) =>
-                      states.contains(WidgetState.selected)
-                      ? null
-                      : Settings.colorSet.secondary,
-                ),
-                value: value,
-                onChanged: (final bool value) {
-                  setState(() {
-                    onChanged(value);
-                    this.value = value;
-                  });
-                },
               ),
             ),
           ),